+++ src/ir_def/construct.ml
@@ -246,10 +246,6 @@ let tagE i e =
 
 let dec_eff dec = match dec.it with
   | LetD (_, e) | VarD (_, _, e) -> eff e
-  | RefD (_, _, le) ->
-    match le.it with
-    | DotLE (e, _) -> eff e
-    | _ -> assert false (*FIXME*)
 
 let rec simpl_decs decs = List.concat_map simpl_dec decs
 and simpl_dec dec = match dec.it with
@@ -544,11 +540,6 @@ let varD x exp =
   let t = typ_of_var x in
   assert (T.is_mut t);
   VarD (id_of_var x, T.as_immut t, exp) @@ no_region
-
-let refD x lexp =
-  let t = typ_of_var x in
-  assert (T.is_mut t);
-  RefD (id_of_var x, t, lexp) @@ no_region
 
 let expD exp =
   let pat = { it = WildP; at = exp.at; note = exp.note.Note.typ } in