+++ src/ir_passes/const.ml
@@ -187,7 +187,7 @@ and gather_dec lvl scope dec : env =
   let mk_info const = { loc_known = lvl = TopLvl; const } in
   let ok = match dec.it with
   | LetD (p, _) -> Ir_utils.is_irrefutable p
-  | VarD _ | RefD _ -> false
+  | VarD _ -> false
   in
   M.fold (fun v _ scope ->
     if ok
@@ -206,8 +206,6 @@ and check_dec lvl env dec : Lbool.t = match dec.it with
     lb
   | VarD (_, _, e) | LetD (_, e) ->
     exp_ lvl env e;
-    surely_false
-  | RefD (_, _, _) ->
     surely_false
 
 and check_decs lvl env ds : Lbool.t =